event/requirements.txt
openlineage-python
numpy
acryl-datahub